Add the node user to the 'dialout' group
If the parent system puts /dev/ttyX in the dialout group then this allows the node user to access the serial ports.

Note that (in my experience) this is not a universal solution as different host systems/distros can put tty's in other groups or have different GID's for dialout. Its probably better to sort this out at runtime, i.e -e SERIALGID=20